UNTESTED! Cleaning qmalloc.
Removed draw related functions.

Removed malloc from klibc.json, multiple definition error.

Now crashes building.

+++ sys/src/9/port/lib.h
@@ -66,16 +66,16 @@ extern	char*	utfrune(char*, int32_t);
 /*
  * malloc
  */
-extern	void*	malloc(uint32_t);
+extern	void*	malloc(size_t);
 extern	void*	mallocz(uint32_t, int);
 extern	void	free(void*);
 extern	uint32_t	msize(void*);
 extern	void*	mallocalign(uint32_t, uint32_t, int32_t, uint32_t);
-extern	void	setmalloctag(void*, uint32_t);
-extern	void	setrealloctag(void*, uint32_t);
-extern	uint32_t	getmalloctag(void*);
-extern	uint32_t	getrealloctag(void*);
-extern	void*	realloc(void *, uint32_t);
+extern	void	setmalloctag(void*, uintptr_t);
+extern	void	setrealloctag(void*, uintptr_t);
+extern	uintptr_t	getmalloctag(void*);
+extern	uintptr_t	getrealloctag(void*);
+extern	void*	realloc(void *, size_t);
 
 /*
  * print routines
